fix: open default diagram after deleting current diagram (#350)
